Move Input::Keys into Input::Key.
