+++ lib/vedeu/input/all.rb
@@ -13,7 +13,6 @@ require 'vedeu/input/input'
 require 'vedeu/input/key'
 require 'vedeu/input/keymap'
 require 'vedeu/input/keymaps'
-require 'vedeu/input/keys'
 require 'vedeu/input/mapper'
 require 'vedeu/input/store'
 require 'vedeu/input/translator'
