--- lib/vedeu/input/keys.rb
@@ -1,13 +1,0 @@
-module Vedeu
-
-  module Input
-
-    # A collection of {Vedeu::Input::Key} instances.
-    #
-    class Keys < Vedeu::Repositories::Collection
-
-    end # Keys
-
-  end # Input
-
-end # Vedeu
