+++ test/lib/vedeu/input/key_test.rb
@@ -4,6 +4,19 @@ module Vedeu
 
   module Input
 
+    describe Keys do
+
+      let(:described) { Vedeu::Input::Keys }
+      let(:instance)  { described.new }
+
+      it { described.superclass.must_equal(Vedeu::Repositories::Collection) }
+
+      describe '#initialize' do
+        it { instance.must_be_instance_of(described) }
+      end
+
+    end # Keys
+
     describe Key do
 
       let(:described) { Vedeu::Input::Key }
